+++ cux_ship/test/json_output_test.dart
@@ -367,6 +367,8 @@ void main() {
           'releaseTypeRaw',
           'copyright',
           'editable',
+          'buildNumber',
+          'buildNumberAsInt',
           'display',
         },
       );